backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
celery[redis]==5.4.0
redis==5.1.1
boto3==1.35.0
botocore==1.35.0
librosa==0.10.2
soundfile==0.12.1
numpy==1.26.4
scipy==1.13.1
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
websockets==13.1
aiofiles==24.1.0
httpx==0.27.2
ffmpeg-python==0.2.0
xmltodict==0.13.0
